+++ changelog.md
@@ -12,6 +12,7 @@
   - Input parser:
     - Add support to expression system for string comparison
     - Add support to expression system for string concat using + not just concat function
+    - Add user variable support to expression system, can access in scripting using client.variables['NAME'] or client.variables.NAME
   - [Preferences](docs/preferences.md)
     - Add `Watch for profile changes` when enabled will watch for profile changes
     - Add `On profile change do` what to do when a profile is changed when `Watch for profile changes` enabled